config: add docker-compose file to be able to quickly deploy the application
create a docker-compose configration to deploy the nodejs app and a
postgresql service. add configs for networks and open the port for the
nodejs app (5000 for app, 9042 if you want to enable the debugger)
and the database (for debugging, if it is needed).
add a docker volume to persist the database.