config: add a Makefile to run and manage docker environment
create a Makefile inspired by gazr normalization to be able to easily
deploy and manage the docker environment.
this config let you launch basic docker-compose command to run and
manage the services inside containers.
add a section to the USER_GUIDE to explain how to use the Makefile
commands to run your environment.

Fixes: https://github.com/MLH/mlh-hackathon-nodejs-starter/issues/3
Refs: https://gazr.io/